selesai perbaiki produk

+++ app/Http/Controllers/AdminController.php
@@ -44,22 +44,32 @@ class AdminController extends Controller
     public function editproduk(Request $request){
 
         $produk = produk::find($request->id);
-        File::delete('data_file/produk/'.$produk->foto);
-
-
+        
         $file = $request->file('file');
+        if($file == null){
+            DB::table('produks')->where('id_produk', $request->id)->update([
+                'nama_produk'=>$request->nama,
+                'harga'=>$request->harga,
+                'deskripsi'=>$request->deskripsi,
+                'kategori'=>$request->kategori,
+            ]);
+        }else{
+            File::delete('data_file/produk/'.$produk->foto);
+        
         // dd($file);
-        $nama_file = $request->nama.".".$file->getClientOriginalExtension();
-        $tujuan_upload = 'data_file/produk';
-        $file->move($tujuan_upload,$nama_file);
-        DB::table('produks')->where('id_produk', $request->id)->update([
-            'nama_produk'=>$request->nama,
-            'harga'=>$request->harga,
-            'file'=>$nama_file,
-            'deskripsi'=>$request->deskripsi,
-            'kategori'=>$request->kategori,
-        ]);
+            $nama_file = $request->nama.".".$file->getClientOriginalExtension();
+            $tujuan_upload = 'data_file/produk';
+            $file->move($tujuan_upload,$nama_file);
+            DB::table('produks')->where('id_produk', $request->id)->update([
+                'nama_produk'=>$request->nama,
+                'harga'=>$request->harga,
+                'file'=>$nama_file,
+                'deskripsi'=>$request->deskripsi,
+                'kategori'=>$request->kategori,
+            ]);
 
+            
+        }
         return redirect()->back();
         // $pegawai = pegawai::where('id_pegawai' ,$id )->get();
         // dd($pegawai);
@@ -95,20 +105,21 @@ class AdminController extends Controller
     public function editpegawai(Request $request){
         
         $pegawai = pegawai::find($request->id_pegawai);
-        File::delete('data_file/pegawai/'.$pegawai->foto);
-        // dd( $request->nama);
-        $file = $request->file('file');
-        $nama_file = $request->nama.".".$file->getClientOriginalExtension();
-        $tujuan_upload = 'data_file/pegawai';
-        $file->move($tujuan_upload,$nama_file);
+        dd($request->id_pegawai);
+        // File::delete('data_file/pegawai/'.$pegawai->foto);
+        // // dd( $request->nama);
+        // $file = $request->file('file');
+        // $nama_file = $request->nama.".".$file->getClientOriginalExtension();
+        // $tujuan_upload = 'data_file/pegawai';
+        // $file->move($tujuan_upload,$nama_file);
         
-        DB::table('pegawais')->where('id_pegawai', $request->id_pegawai)->update([
-            'nama_pegawai'=>$request->nama,
-            'alamat'=>$request->alamat,
-            'foto'=>$nama_file,
-            'no_hp'=>$request->no_hp
-        ]);
+        // DB::table('pegawais')->where('id_pegawai', $request->id_pegawai)->update([
+        //     'nama_pegawai'=>$request->nama,
+        //     'alamat'=>$request->alamat,
+        //     'foto'=>$nama_file,
+        //     'no_hp'=>$request->no_hp
+        // ]);
         
-        return redirect()->back();
+        // return redirect()->back();
     }  
 }
